+++ NEWS.md
@@ -1,6 +1,8 @@
 Version 0.14.2-dev
 2013-11-DD
 
+ - Fix support for exponential notation in filter values (#311)
+
 Version 0.14.1
 2013-10-31